modify：用户信息列表返回格式

+++ src/main/java/com/gohead/core/dao/UserInfoMapper.java
@@ -37,5 +37,5 @@ public interface UserInfoMapper {
      * @param map
      * @return
      */
-    public Long getTotalUserInfo(Map<String, Object> map);
+    public int getTotalUserInfo(Map<String, Object> map);
 }
